+++ archetypes/api/_APPNAME_/_APPNAME_-api/src/main/java/applica/_APPNAME_/api/utils/InvalidDataException.java
@@ -3,5 +3,8 @@ package applica._APPNAME_.api.utils;
 /**
  * Created by bimbobruno on 27/01/2017.
  */
-public class InvalidDataException {
+public class InvalidDataException extends Exception {
+
+    public InvalidDataException(String data) {
+    }
 }