+++ archetypes/client/app/js/stores.js
@@ -160,7 +160,7 @@ export const MenuStore = aj.createStore(MENU, (state = {}, action) => {
             return _.assign(state, {menu: action.menu})
 
         case actions.SET_ACTIVE_MENU_ITEM:
-            return _.assign(state, {menu: walk(state.menu, "children", i => { i.active = i == action.item})})
+            return _.assign(state, {menu: walk(state.menu, "children", i => { console.log(i); console.log(action.item), console.log(action.item == i); i.active = (i == action.item)})})
 
     }